Expand the list of libraries that can be provided by the host
nvidia-container-toolkit will provide all of these libraries to a
container if the NVIDIA_DRIVER_CAPABILITIES=all environment variable is
set. To avoid conflicts with the host let's not generate provides for
any of them.

The list of libraries was generated by installing
nvidia-container-toolkit 1.17.8-1 on an Ubuntu 24.04 system with an
NVIDIA GPU and then running the Chainguard bash docker container with
`-e NVIDIA_DRIVER_CAPABILITIES=all --gpus all` and checking /usr/lib/
for all libraries with the same version number as the NVIDIA drivers
installed on the host.