ci: make the vuln scan simpler
We were fetching advisory data as an input to a filtered scan, similar to what we'd do in a packages repo, but that's really unnecessary here since we're not concerned with how many vulns are in the output, we just care that the scanning operation itself doesn't blow up. For the same reason, we can remove the strict --require-zero flag that would only serve as a nuisance for melange changes.

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>