Propagate environment between uses and pipelines
If you specify the environment of a uses pipeline it will
be ignored when the pipeline is loaded.

You can see grpc-1.76 doing this here: But I kept seeing the python binaries downloading contents.

https://github.com/wolfi-dev/os/blob/0c6b8959949e816f3dbce1af2237ad81ee579618/grpc-1.74.yaml#L120

These variables aren't ever actually set the uses pipeline overwrites it. So we need to fix those
places so that the variables are propegated. An easy fix for this when it's needed is to make then
environment variables at the top level with the packages. This also explains some additional strange
behavior we've had trying to override variables for individual steps.