Download alpine kernel for local testing
For some time you haven't been able to `make test-e2e` in melange.
To fix this we'll download the alpine kernel to use for testing just like the CI job does.

I'm debugging this on a arm machine too and amd64 stuff is too slow so enable the tests to be run in the same arch as the host by default.