qemu_runner: Identify as qemu intentionally
Use those in the tests that require them. The current check doesn't work on
qemu emulating either aarch64 or x86_64 on my mac and when I put the docker
runner in qemu mode the path is also not there. If we want to test this
functionality we should just test it. If we don't want to leak or env into
the build we can key this of an env variable or flag to melange.