qemu runner: rename SSH connection info
Rename the variables around handling SSH connections with the build
guest to make it clearer which ones are intended to handle the build
actions and which ones handle controlling the guest, and what has
privileges.

Essentially, the following transitions:

  SSHWorkspaceAddress -> SSHControlAddress
  sshWorkspacePort -> sshControlPort

  SSHClient -> SSHBuildClient
  WorkspaceClient -> SSHControlClient

This is purely a renaming change, there should be no functional
changes as a result.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>