tests: add checks for run-as root and non-root users
The QEMU runner broke the ability to have non-root builds, and the
bubblewrap runner fails when the group ID does not match the user ID
of the run-as user.

v2: add tests to ensure numeric versions of user ids as specified in the
    `run-as:` statement correctly function as the intended user.
v3: add check-accounts pipeline and use that in the accounts tests.
v4: add checks that tests for subpackages can run as different users
    than the parent package

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>
MR: https://github.com/chainguard-dev/melange/pull/2138