bubblewrap: when running as custom UID, use corrsponding GID
When running a build with a non-standard UID, lookup and use the
effective GID that corresponds to that User/UID as configured. Without
this, under bubblewrap it would default to using the UID of the runas
user, which is not necessarily a valid group.

v2: fall back to using GID == UID if GID is not specified, and fix up
    unit tests to test for both situations.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>