pipelines/go/build: Add ignore-untracked-files to prevent some +dirty builds (#2146)
Right now we have some files that are added to the build env
that may cause binaries to be built with +dirty in the verison when
inappropriate. This adds and option to ignore those loose files.

It doesn't cover modifications to tracked files and commits will still
be appended to the version.