Propagate environment between uses and pipelines (#2139)
If you specify the environment of a uses pipeline it will
be ignored when the pipeline is loaded.

You can see grpc-1.76 doing this here:

https://github.com/wolfi-dev/os/blob/0c6b8959949e816f3dbce1af2237ad81ee579618/grpc-1.74.yaml#L120

These variables aren't ever actually set, the uses pipeline overwrites them. So we need to fix those
so that the variables are propagated. An easy fix for this when it's needed is to make then
environment variables at the top level. This also explains some additional strange
behavior we've had trying to override variables for individual steps for
compiler options like the openssf options.