qemu runner: report kernel version (#2132)
Capture and report the kernel version signature in the qemu guest;
this is useful auditing information to know what kernel is in the
environment, especially for diagnosing build or test failures that are
due to kernel config settings not being correct.

The result is a log message that looks like:

```
2025/08/18 21:51:00 INFO qemu: running kernel version: 6.15.8-r1-qemu-generic #Chainguard SMP PREEMPT_DYNAMIC Mon Jul 28 23:48:27 UTC 2025
```

This is the emitted output of `uname -vr` in the guest.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>