sca: fix SCA for go-fips-1.25 (#2153)
Ensure that Melange SCA generates depends for packages that use the
new go-msft-1.25 toolchain.

Tried to create a test case for this, but failed in:
- https://github.com/chainguard-dev/melange/pull/2152

It would require enterprise-packages access which is not great.

Separately test cases are available as, currently:
- 1.24 works https://github.com/chainguard-dev/enterprise-packages/pull/33090
- 1.25 doesn't https://github.com/chainguard-dev/enterprise-packages/pull/33091

And with this PR both should start working again.

Suggestions on how to make test cases work are welcomed to catch this.

My plan is to add SCA check to go-msft packages themselves, such that it breaks when there are changes to build-info in the future:
- https://github.com/chainguard-dev/enterprise-packages/pull/32952

Also note that libssl.so is unused, and only libcrypto is dlopened.