qemu runner: bump level of logging for initial connection failure
We are seeing some build failures in production where melange does not
appear to be able to get the openssh protocol version string to know
that the ssh service has successfully started, though log messages
indicate that the ssh daemon has started enough to notice melange's
attempts to make initial contact.

This can be reverted once we've identified what's causing the issue.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>