fix(user accounts): don't crash if user without gid is specified
Under the bubblewrap runner, if a user was added to run
the build under, melange would always set the gid to be
the same as the uid of the user, including in the case when
no gid was given for the user account. However, as part of
https://github.com/chainguard-dev/melange/pull/2138 , commit
a3612ee ("bubblewrap: when running as custom UID, use corrsponding
GID") modified the bubblewrap runner to attempt to look up the
primary gid of the user the build was set to run as. However,
that lookup did not take into account the possibility that no
gid was configured, and thus would cause melange to segfault,
as identified in the followup comment to the above PR:
https://github.com/chainguard-dev/melange/pull/2138#issuecomment-3275304078

Fix this by falling back to the old behavior when no gid is specified
for the user account the build is running as, and add an e2e test
yaml for this situation. Both the bubblewrap and qemu runners now
use the same behavior in this situation.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>