Qemu runner: debug initial connection (#2154)
* qemu runner: bump level of logging for initial connection failure

We are seeing some build failures in production where melange does not
appear to be able to get the openssh protocol version string to know
that the ssh service has successfully started, though log messages
indicate that the ssh daemon has started enough to notice melange's
attempts to make initial contact.

This can be reverted once we've identified what's causing the issue.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>

* qemu runner: increase inital connection timeout

We are seeing some build failures in production where melange does not
appear to be able to get the openssh protocol version string to know
that the ssh service has successfully started, though log messages
indicate that the ssh daemon has started enough to notice melange's
attempts to make initial contact.

Bump the Dial timeout from 50ms to 150ms to see if resource contention
is causing the 50ms timeout to be exceeded.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>

---------

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>