R/build pipeline: stop installing into system dirs
The melange R/build pipeline would invoke the R build install process in
a way that caused the R package being built to be installed into
/usr/lib/R/library and then moved into the melange staging directory.
This is both inefficient and breaks building packages with a non-root
user that does not have the privilege to write to system directories.

Fix this by adjusting the install.packages Rscript call to point to the
melange staging area directly.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>