R build pipeline improvements (#2160)
* R/build pipeline: stop installing into system dirs

The melange R/build pipeline would invoke the R build install process in
a way that caused the R package being built to be installed into
/usr/lib/R/library and then moved into the melange staging directory.
This is both inefficient and breaks building packages with a non-root
user that does not have the privilege to write to system directories.

Fix this by adjusting the install.packages Rscript call to point to the
melange staging area directly.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>

* R/build pipeline: add R-doc build dependency

R library packages issue warnings at build time when installing their
stub html help documents:

    *** installing help indices
    Warning in file.copy(file.path(R.home("doc"), "html", "R.css"), outman) :
    problem copying /usr/share/doc/R/html/R.css to /home/build/melange-out/Rcpp/usr/lib/R/library/00LOCK-build/00new/Rcpp/html/R.css: No such file or directory

This is because the R.css file is located in the R-doc package.

Fix this by adding the R-doc package as a default build-time dependency
for packages using the R/build pipeline.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>

---------

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>