qemu runner: open permissions on melange cache upper overlay
As part of supporting non-root builds in the qemu runner, make the
upper portion of the melange_cache mount overlay be world writable with
the sticky bit set to act similar to a tmpdir. This way non-root builds
can write to the overlay. Also add a check that this directory is
writable during builds in the e2e account tests.

(Because of the overlay, these permissions and writes will not write
through to the underlying mounted filesystem from the host).

Also reformat the injected sequence of commands to make them more
human readable.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>