Add new source subcommand This subcommand pulls the source a build config. Currently it only supports git-checkout but in the future it should suppport more functionality.
The purpose of this is to make interacting with upstream source and
patches easier.  I'd like a workflow like `melange source -o sources/
package.yaml` to setup a git tree with patches applied and another
command that can do the revese.

There's something similiar in source.go for license checks. That version
runs the same shell script that the pipeline has.  This command is meant
for setting up an environment that can be used to develope against the
upstream sources. I'm not interested shallow cloning or keeping a
consistent commiter id or anything here. I'm realy just trying to make
it easy to develop/debug packages using the upstream source.