fetch pipeline: don't adjust ownership when untarring as root
When trying to compensate for defaulting to perform qemu builds as
non-root by default, some packages perform operations that need to be
root privileges in qemu, like setting filesystem capabilities. However,
running as root can cause a failure if the build is performed under the bubblewrap
runner when using the fetch pipeline to pull and untar a tarball that
has contents that are owned by non-root uids; see the failed builds in
https://github.com/chainguard-dev/enterprise-packages/pull/36668
for example:

```
2025-10-22 18:06:40.247 tar: nginx-1.25.5/man/nginx.8: Cannot change ownership to uid 502, gid 20: Invalid argument
2025-10-22 18:06:40.247 tar: nginx-1.25.5: Cannot change ownership to uid 502, gid 20: Invalid argument
2025-10-22 18:06:40.247 tar: Exiting with failure status due to previous errors
```

This is a similar problem that was addressed for the git fetch pipeline
in https://github.com/chainguard-dev/melange/pull/1893

Fix this here by passing `--no-same-owner` to tar in the fetch pipeline.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>