qemu runner: make virtfs 9p bindings read-only v2 (#2155)
* qemu: virtfs 9p bindings read-only, fix copy permissions, take 2

The QEMU runner uses virtfs 9p mount bindings to inject some artifacts
into the build environment, but unlike the docker and bubblewrap
runners, does not use any bind mounts to extract out artifacts to
produce apks; instead tar over ssh is used for artifact extraction.

Thus, to reduce risk of a malicious build being able to write artifacts
outside of the QEMU guest, make the virtfs 9p bindings be read-only.
(There has also has been at least one vulnerability the virtfs 9p
implementation that allowed writes outside of the specified path.)

As part of this, start the process of moving the defaultshare 9p
mount into melange directly, rather than relying on microvm-init to do
it for us. However, because it currently does do it for us and also
makes permissions changes to the mount, we need to temporarily
compensate for this by mounting the filesystem if it's not been already
mounted or remounting it readonly if microvm-init has done it for us.

Also, because of the elimination in microvm-init of adjustment of the
permissions on the contents of the defaultshare before copying said
contents into the /home/build directory, this breaks non-root builds
where the uid of guest builder does not match the uid of the process
outside the guest invoking melange+qemu (9p's mapped permissions map
the external uid running qemu as the uid owner of the contents of the
share in the guest). Fix this by adjusting the ownership of the files
after the copy to be the user that the build will be performed under.

Finally, unmount the defaultshare 9p mount entirely after copying
the contents of it into the build directory, if the umount command is in
the guest's control environment.

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>

* Update pkg/container/qemu_runner.go

Co-authored-by: Evan Gibler <20933572+egibs@users.noreply.github.com>
Signed-off-by: Steve Beattie <steve+github@nxnw.org>

* sbom_group_test.go: fix unrelated import linter issue

fixes: 601baee8 ("Update pkg/container/qemu_runner.go")
Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>

---------

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>
Signed-off-by: Steve Beattie <steve+github@nxnw.org>
Co-authored-by: Evan Gibler <20933572+egibs@users.noreply.github.com>