pkg: Emit appCpe stanza (#2280)
Use CPEString() function, to generated appCpe when a package Vendor is set.

This is to comply with the new revision of UAPI Group Specification
and to hint this information to SBOM generators and CVE scanners (for
example grype has implemented support for this).

Note it is quite odd to have "*" value for the CPE part (or type), as
there are currently only three valid values ("a" application, "h"
hardware, and "o" os). hence init the default CPE to "a" application,
and validate that only these three values are acceptable.

There are no direct tests for this, but workplace population is still
successful. And ultimately it matters that gcc is able to process the
new link file and emit correct metadata. I have validated this by
manually building zip (no CPE data) & giflib (with CPE data), and both
build and link correctly.

Once grype/syft support is stable and out in the next release, we
could look into making e2e integration test => come up with a custom
CPE and check that produced package/binary has the declared appCpe in
the syft sbom result.